Refactor webview host handler modules (oh-tab-17cv)
